collapse_de_pages

+++ couches/control/supr.php
@@ -6,35 +6,36 @@ if (!isset($_SESSION)) {
   header("Location: signup&login_form.php");
 }
 
+if(isset($_GET['noemp'])){
 
-$empService = new EmployeService;
-try{
-  $sups = $empService->selectAllSupsNum();
-
-}catch(ServiceException $e){
-
-}
-$sups_1d = [];
-foreach ($sups as $sup) {
-  $sups_1d[] = $sup['noemp'];
-}
-if (!in_array($_GET['noemp'], $sups_1d)) {
-  $noemp = $_GET['noemp'];
+  $empService = new EmployeService;
   try{
-
-    $deleted = $empService->deleteEmp($noemp);
+    $sups = $empService->selectAllSupsNum();
+  
   }catch(ServiceException $e){
-
+  
   }
-
-
-
-  header("Location: tableau-connecte.php?suppression=succes");
-  // echo 'here';
-
-} else {
-  echo "impossible de supprimer un supérieur";
-}
+  $sups_1d = [];
+  foreach ($sups as $sup) {
+    $sups_1d[] = $sup['noemp'];
+  }
+  if (!in_array($_GET['noemp'], $sups_1d)) {
+    $noemp = $_GET['noemp'];
+    try{
+  
+      $deleted = $empService->deleteEmp($noemp);
+    }catch(ServiceException $e){
+  
+    }
+  
+  
+  
+    header("Location: tableau-connecte.php?suppression=succes");
+    // echo 'here';
+  
+  } else {
+    echo "impossible de supprimer un supérieur";
+  }
 
 
 ?>
